Refactoring similar rules to avoid code duplication.
