+++ java-checks/src/main/java/org/sonar/java/checks/security/DataHashingCheck.java
@@ -19,215 +19,9 @@
  */
 package org.sonar.java.checks.security;
 
-import com.google.common.collect.ImmutableMap;
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.HashMap;
-import java.util.Locale;
-import java.util.Map;
-import java.util.Optional;
 import org.sonar.check.Rule;
-import org.sonar.java.checks.StrongCipherAlgorithmCheck;
-import org.sonar.java.checks.helpers.JavaPropertiesHelper;
-import org.sonar.java.checks.methods.AbstractMethodDetection;
-import org.sonar.java.model.ExpressionUtils;
-import org.sonar.java.model.LiteralUtils;
-import org.sonar.plugins.java.api.semantic.MethodMatchers;
-import org.sonar.plugins.java.api.tree.ExpressionTree;
-import org.sonar.plugins.java.api.tree.IdentifierTree;
-import org.sonar.plugins.java.api.tree.LiteralTree;
-import org.sonar.plugins.java.api.tree.MethodInvocationTree;
-import org.sonar.plugins.java.api.tree.NewClassTree;
-import org.sonar.plugins.java.api.tree.Tree;
-
-import static org.sonar.java.checks.security.DataHashingCheck.InsecureAlgorithm.MD2;
-import static org.sonar.java.checks.security.DataHashingCheck.InsecureAlgorithm.MD5;
-import static org.sonar.java.checks.security.DataHashingCheck.InsecureAlgorithm.SHA1;
-import static org.sonar.plugins.java.api.semantic.MethodMatchers.ANY;
+import org.sonar.java.checks.AbstractHashAlgorithmChecker;
 
 @Rule(key = "S4790")
-public class DataHashingCheck extends AbstractMethodDetection {
-  private static final String JAVA_LANG_STRING = "java.lang.String";
-  private static final String GET_INSTANCE = "getInstance";
-  private static final String CONSTRUCTOR = "<init>";
-
-  enum InsecureAlgorithm {
-    MD2, MD4, MD5, MD6, RIPEMD,
-    HAVAL128 {
-      @Override
-      public String toString() {
-        return "HAVAL-128";
-      }
-    },
-    SHA1 {
-      @Override
-      public String toString() {
-        return "SHA-1";
-      }
-    },
-    DSA {
-      @Override
-      boolean match(String algorithm) {
-        // exact match required for DSA, so it doesn't match ECDSA
-        return "DSA".equals(algorithm);
-      }
-    };
-
-    boolean match(String algorithm) {
-      String normalizedName = algorithm.replace("-", "").toLowerCase(Locale.ENGLISH);
-      return normalizedName.contains(name().toLowerCase(Locale.ENGLISH));
-    }
-  }
-
-  private enum DeprecatedSpringPasswordEncoder {
-    MD5("org.springframework.security.authentication.encoding.Md5PasswordEncoder", CONSTRUCTOR),
-    SHA("org.springframework.security.authentication.encoding.ShaPasswordEncoder", CONSTRUCTOR),
-    LDAP("org.springframework.security.crypto.password.LdapShaPasswordEncoder", CONSTRUCTOR),
-    MD4("org.springframework.security.crypto.password.Md4PasswordEncoder", CONSTRUCTOR),
-    MESSAGE_DIGEST("org.springframework.security.crypto.password.MessageDigestPasswordEncoder", CONSTRUCTOR),
-    STANDARD("org.springframework.security.crypto.password.StandardPasswordEncoder", CONSTRUCTOR),
-    NO_OP("org.springframework.security.crypto.password.NoOpPasswordEncoder", GET_INSTANCE);
-
-    private static final String MESSAGE_FORMAT = "Don't rely on %s because it is deprecated and use a stronger hashing algorithm.";
-    protected static final Map<String, String> MESSAGE_PER_CLASS;
-
-    static {
-      MESSAGE_PER_CLASS = new HashMap<>();
-      MESSAGE_PER_CLASS.put(MD5.classFqn, "Use a stronger hashing algorithm than MD5.");
-      MESSAGE_PER_CLASS.put(SHA.classFqn, "Don't rely on " + SHA.className + " because it is deprecated.");
-      MESSAGE_PER_CLASS.put(LDAP.classFqn, String.format(MESSAGE_FORMAT, LDAP.className));
-      MESSAGE_PER_CLASS.put(MD4.classFqn, String.format(MESSAGE_FORMAT, MD4.className));
-      MESSAGE_PER_CLASS.put(MESSAGE_DIGEST.classFqn, String.format(MESSAGE_FORMAT, MESSAGE_DIGEST.className));
-      MESSAGE_PER_CLASS.put(NO_OP.classFqn, "Use a stronger hashing algorithm than this fake one.");
-      MESSAGE_PER_CLASS.put(STANDARD.classFqn, "Use a stronger hashing algorithm.");
-    }
-
-    private final String classFqn;
-    private final String methodName;
-    private final String className;
-
-    DeprecatedSpringPasswordEncoder(String fqn, String methodName) {
-      this.classFqn = fqn;
-      this.methodName = methodName;
-      String[] fqnParts = fqn.split("\\.");
-      this.className = fqnParts[fqnParts.length - 1];
-    }
-  }
-
-  /**
-   * These APIs have static getInstance method to get an implementation of some crypto algorithm.
-   * javax.crypto.Cipher is missing from this list, because it is covered by rule S5547 {@link StrongCipherAlgorithmCheck}
-   * Details can be found here <a href="http://docs.oracle.com/javase/8/docs/technotes/guides/security/StandardNames.html">Security Standard Names</a>
-   */
-  private static final String[] CRYPTO_APIS = {
-    "java.security.AlgorithmParameters",
-    "java.security.AlgorithmParameterGenerator",
-    "java.security.MessageDigest",
-    "java.security.KeyFactory",
-    "java.security.KeyPairGenerator",
-    "java.security.Signature",
-    "javax.crypto.Mac",
-    "javax.crypto.KeyGenerator"
-  };
-
-  private static final Map<String, InsecureAlgorithm> ALGORITHM_BY_METHOD_NAME = ImmutableMap.<String, InsecureAlgorithm>builder()
-    .put("getMd2Digest", MD2)
-    .put("getMd5Digest", MD5)
-    .put("getShaDigest", SHA1)
-    .put("getSha1Digest", SHA1)
-    .put("md2", MD2)
-    .put("md2Hex", MD2)
-    .put("md5", MD5)
-    .put("md5Hex", MD5)
-    .put("sha1", SHA1)
-    .put("sha1Hex", SHA1)
-    .put("sha", SHA1)
-    .put("shaHex", SHA1)
-    .put("md5Digest", MD5)
-    .put("md5DigestAsHex", MD5)
-    .put("appendMd5DigestAsHex", MD5)
-    .build();
-
-  @Override
-  protected MethodMatchers getMethodInvocationMatchers() {
-    ArrayList<MethodMatchers> matchers = new ArrayList<>();
-    matchers
-      .add(MethodMatchers.create()
-        .ofTypes("org.apache.commons.codec.digest.DigestUtils")
-        .names("getDigest")
-        .addParametersMatcher(JAVA_LANG_STRING)
-        .build());
-
-    matchers
-      .add(MethodMatchers.create()
-      .ofTypes("org.apache.commons.codec.digest.DigestUtils")
-      .name(ALGORITHM_BY_METHOD_NAME::containsKey)
-      .withAnyParameters()
-      .build());
-
-    matchers
-      .add(MethodMatchers.create()
-        .ofTypes(CRYPTO_APIS)
-        .names(GET_INSTANCE)
-        .addParametersMatcher(JAVA_LANG_STRING)
-        .addParametersMatcher(JAVA_LANG_STRING, ANY)
-        .build());
-
-    matchers
-      .add(MethodMatchers.create()
-        .ofTypes("org.springframework.util.DigestUtils")
-        .names("appendMd5DigestAsHex", "md5Digest", "md5DigestAsHex")
-        .withAnyParameters()
-        .build());
-
-    for (DeprecatedSpringPasswordEncoder pe : DeprecatedSpringPasswordEncoder.values()) {
-      matchers.add(MethodMatchers.create().ofTypes(pe.classFqn).names(pe.methodName).withAnyParameters().build());
-    }
-
-    matchers.add(MethodMatchers.create()
-      .ofTypes("com.google.common.hash.Hashing")
-      .names("md5", "sha1")
-      .addWithoutParametersMatcher().build());
-
-    return MethodMatchers.or(matchers);
-  }
-
-  @Override
-  protected void onMethodInvocationFound(MethodInvocationTree mit) {
-    IdentifierTree methodName = ExpressionUtils.methodName(mit);
-    String message = DeprecatedSpringPasswordEncoder.MESSAGE_PER_CLASS.get(methodName.symbol().owner().type().fullyQualifiedName());
-    if (message != null) {
-      reportIssue(methodName, message);
-      return;
-    }
-    InsecureAlgorithm algorithm = ALGORITHM_BY_METHOD_NAME.get(methodName.name());
-    if (algorithm == null) {
-      algorithm = algorithm(mit.arguments().get(0)).orElse(null);
-    }
-    if (algorithm != null) {
-      reportIssue(methodName, "Use a stronger hashing algorithm than " + algorithm.toString() + ".");
-    }
-  }
-
-  @Override
-  protected void onConstructorFound(NewClassTree newClassTree) {
-    String message = DeprecatedSpringPasswordEncoder.MESSAGE_PER_CLASS.get(newClassTree.identifier().symbolType().fullyQualifiedName());
-    reportIssue(newClassTree.identifier(), message);
-  }
-
-  private static Optional<InsecureAlgorithm> algorithm(ExpressionTree invocationArgument) {
-    ExpressionTree expectedAlgorithm = invocationArgument;
-    ExpressionTree defaultPropertyValue = JavaPropertiesHelper.retrievedPropertyDefaultValue(invocationArgument);
-    if (defaultPropertyValue != null) {
-      expectedAlgorithm = defaultPropertyValue;
-    }
-    if (expectedAlgorithm.is(Tree.Kind.STRING_LITERAL)) {
-      String algorithmName = LiteralUtils.trimQuotes(((LiteralTree) expectedAlgorithm).value());
-      return Arrays.stream(InsecureAlgorithm.values())
-        .filter(alg -> alg.match(algorithmName))
-        .findFirst();
-    }
-    return Optional.empty();
-  }
-
+public class DataHashingCheck extends AbstractHashAlgorithmChecker {
 }
